Profile more.
If I had to guess, the reason most of the time is spent on copyTask is that it's calling new[] and delete[] a LOT.

Idea: Make a DeadTaskQueue. 8192 limit, deleting tasks sends it to the Dead Task Queue (unless past 8192 limit), copyTask takes from the dead task queue unless it's empty then it creates a new task.