Prepare for knapsack problem
Wikipedia's page on the Knapsack Problem has a solution for the 0-1 knapsack problem.

As it turns out, if you let MAX_DL = infinity and ignore the "two sets covering one series does not add that series value twice" restriction, creating a DL in Mudae Bot is the same as the 0-1 knapsack problem. So I MIGHT be able to use that to perform SOME optimization, not sure what though.