Kill worse tasks
Run a diff checker on the following two files. (I already removed the text about better DL found, yadda yadda)

The first text is the old system. The second text is the new system. I say "run a diff checker" because there WILL be differences, but all of the differences are "there are way less tasks"

First text file:

```Best score in knapsack: 0
With a setDeleteIndex of 1,
the inTaskQueue has 2 tasks

Best score in knapsack: 12621
With a setDeleteIndex of 2,
the inTaskQueue has 4 tasks

Best score in knapsack: 15809
With a setDeleteIndex of 3,
the inTaskQueue has 8 tasks

Best score in knapsack: 17315
With a setDeleteIndex of 4,
the inTaskQueue has 16 tasks

Best score in knapsack: 19786
With a setDeleteIndex of 5,
the inTaskQueue has 31 tasks

Best score in knapsack: 19786
With a setDeleteIndex of 6,
the inTaskQueue has 47 tasks

Best score in knapsack: 19786
With a setDeleteIndex of 7,
the inTaskQueue has 86 tasks

Best score in knapsack: 19786
With a setDeleteIndex of 8,
the inTaskQueue has 128 tasks

Best score in knapsack: 19786
With a setDeleteIndex of 9,
the inTaskQueue has 229 tasks

Best score in knapsack: 19975
With a setDeleteIndex of 10,
the inTaskQueue has 401 tasks

Best score in knapsack: 20505
With a setDeleteIndex of 11,
the inTaskQueue has 664 tasks

Best score in knapsack: 20505
With a setDeleteIndex of 12,
the inTaskQueue has 1153 tasks

Best score in knapsack: 20505
With a setDeleteIndex of 13,
the inTaskQueue has 1925 tasks

Best score in knapsack: 20505
With a setDeleteIndex of 14,
the inTaskQueue has 2917 tasks

Best score in knapsack: 20505
With a setDeleteIndex of 15,
the inTaskQueue has 4930 tasks

Best score in knapsack: 20505
With a setDeleteIndex of 16,
the inTaskQueue has 6688 tasks

Best score in knapsack: 20601
With a setDeleteIndex of 17,
the inTaskQueue has 9893 tasks

Best score in knapsack: 20611
With a setDeleteIndex of 18,
the inTaskQueue has 14727 tasks

Best score in knapsack: 20611
With a setDeleteIndex of 19,
the inTaskQueue has 21607 tasks

Best score in knapsack: 20611
With a setDeleteIndex of 20,
the inTaskQueue has 38026 tasks

Best score in knapsack: 20611
With a setDeleteIndex of 21,
the inTaskQueue has 67121 tasks

Best score in knapsack: 20779
With a setDeleteIndex of 22,
the inTaskQueue has 121471 tasks```

Second text file:

```Best score in knapsack: 0
With a setDeleteIndex of 1,
the inTaskQueue has 2 tasks

Best score in knapsack: 12621
With a setDeleteIndex of 2,
the inTaskQueue has 4 tasks

Best score in knapsack: 15809
With a setDeleteIndex of 3,
the inTaskQueue has 8 tasks

Best score in knapsack: 17315
With a setDeleteIndex of 4,
the inTaskQueue has 15 tasks

Best score in knapsack: 19786
With a setDeleteIndex of 5,
the inTaskQueue has 26 tasks

Best score in knapsack: 19786
With a setDeleteIndex of 6,
the inTaskQueue has 32 tasks

Best score in knapsack: 19786
With a setDeleteIndex of 7,
the inTaskQueue has 50 tasks

Best score in knapsack: 19786
With a setDeleteIndex of 8,
the inTaskQueue has 61 tasks

Best score in knapsack: 19786
With a setDeleteIndex of 9,
the inTaskQueue has 78 tasks

Best score in knapsack: 19975
With a setDeleteIndex of 10,
the inTaskQueue has 87 tasks

Best score in knapsack: 20505
With a setDeleteIndex of 11,
the inTaskQueue has 87 tasks

Best score in knapsack: 20505
With a setDeleteIndex of 12,
the inTaskQueue has 106 tasks

Best score in knapsack: 20505
With a setDeleteIndex of 13,
the inTaskQueue has 121 tasks

Best score in knapsack: 20505
With a setDeleteIndex of 14,
the inTaskQueue has 124 tasks

Best score in knapsack: 20505
With a setDeleteIndex of 15,
the inTaskQueue has 140 tasks

Best score in knapsack: 20505
With a setDeleteIndex of 16,
the inTaskQueue has 136 tasks

Best score in knapsack: 20601
With a setDeleteIndex of 17,
the inTaskQueue has 150 tasks

Best score in knapsack: 20611
With a setDeleteIndex of 18,
the inTaskQueue has 144 tasks

Best score in knapsack: 20611
With a setDeleteIndex of 19,
the inTaskQueue has 191 tasks

Best score in knapsack: 20611
With a setDeleteIndex of 20,
the inTaskQueue has 175 tasks

Best score in knapsack: 20611
With a setDeleteIndex of 21,
the inTaskQueue has 247 tasks

Best score in knapsack: 20779
With a setDeleteIndex of 22,
the inTaskQueue has 331 tasks```